FIX predict function
helper functions where discarded and predict function now works standalone.
