+++ model_bridging/model_bridging/predict.py
@@ -1,19 +1,86 @@
 def predict(**kwargs):
     import pandas as pd
     import numpy as np
-    from model_bridging.helpers import (
-        tokenize_pd_code,
-        get_date_diff,
-        add_emp_tenure_to_df,
-        get_bucket_and_key_from_s3_uri,
-        download_model_from_s3
-    )
+    import boto3
+    import joblib
+    import tempfile
+    from typing import List
     from nltk.corpus import stopwords
     from nltk.stem.snowball import SnowballStemmer
     from nltk.tokenize import WhitespaceTokenizer
     import json
     import nltk
+
     nltk.download("stopwords")
+
+    def get_bucket_and_key_from_s3_uri(uri):
+        bucket, key = uri.split("/", 2)[-1].split("/", 1)
+        return bucket, key
+
+    def tokenize_pd_code(df):
+
+        df[["pd_code_1", "pd_code_2"]] = pd.DataFrame(
+            df["PrimaryDiagnosisCode"].fillna("na.na").str.split(".", expand=True)
+        )
+        return df.drop("PrimaryDiagnosisCode", axis=1)
+
+    def get_date_diff(col1, col2, interval):
+        """difference between dates specified by the interval in ['Y','D','M']
+        col1 and col2 are date colummns and col2 > col1"""
+
+        return (col2 - col1) / np.timedelta64(1, interval)
+
+    def add_emp_tenure_to_df(df):
+        """
+        returns a df with employment tenure column appeneded
+        """
+        df["emp_tenure"] = get_date_diff(
+            df["InsuredHireDate"], df["LossDate"], interval="D"
+        )
+
+        return df.drop("InsuredHireDate", axis=1)
+
+    # joblib needs this class
+    class CategoricalGrouper:
+
+        """we are fitting the categorical grouper on a reference data
+        using min_freq, which means all the low-frequency categories are
+        pushed into a garbage or OOV category"""
+
+        def __init__(self, min_freq):
+            self.min_freq = min_freq
+            self.is_fitted = False
+
+        def fit(self, X, categorical: List):
+            _non_oov_dict = dict()
+            for col in categorical:
+                _df = pd.DataFrame(X[col].value_counts()).reset_index(drop=False)
+                _df.columns = ["index", "counts"]
+                non_oov = list(_df.loc[_df["counts"] >= self.min_freq, "index"])
+                _non_oov_dict[col] = non_oov
+            #             X.loc[~X[col].isin(non_oov), col] = 'OOV'
+            self._non_oov_dict = _non_oov_dict
+            self.is_fitted = True
+            return None
+
+        def transform(self, X, categorical: List):
+            if self.is_fitted:
+                for col in categorical:
+                    non_oov = self._non_oov_dict[col]
+                    X.loc[~X[col].isin(non_oov), col] = "OOV"
+
+                return X
+            else:
+                raise ValueError(
+                    "The categorical grouper needs to be fit before calling .transform()"
+                )
+
+    def download_model_from_s3(bucket_name, key):
+        bucket = boto3.resource("s3").Bucket(bucket_name)
+        with tempfile.NamedTemporaryFile() as fp:
+            bucket.download_fileobj(key, fp)
+            loaded_model = joblib.load(fp.name)
+        return loaded_model
 
     def stem_text(text):
         return " ".join([stemmer.stem(w) for w in w_tokenizer.tokenize(text)])
@@ -48,11 +115,11 @@ def predict(**kwargs):
 
     for artifact in kwargs.get("artifact"):
         if artifact.get("dataName") == "combined_artifacts":
-            model_bucket, model_key = get_bucket_and_key_from_s3_uri(artifact.get("dataValue"))
+            model_bucket, model_key = get_bucket_and_key_from_s3_uri(
+                artifact.get("dataValue")
+            )
 
     artifacts = download_model_from_s3(model_bucket, model_key)
-
-
 
     # Unpacking artifacts from the joblib object
     model = artifacts.get("model")
@@ -60,9 +127,7 @@ def predict(**kwargs):
     categorical_grouper = artifacts.get("categorical_grouper")
     train_template = artifacts.get("train_template")
 
-
     input_data = pd.DataFrame([kwargs.get("inputs").get("claim")])
-
 
     date_cols = [x for x in input_data.columns if "date" in x.lower()]
     for col in date_cols:
@@ -167,8 +232,18 @@ def predict(**kwargs):
     payload_json = json.loads(pred_payload.to_json(orient="records"))[0]
     claim_number = payload_json["ClaimNumber"]
     return [
-        {"inputDataSource": f"{claim_number}:0", "entityId": claim_number,
-         "predictedResult": [{'claimNumber':claim_number,'predictedProbability':payload_json['predicted_probability'],'predicted_bridged_ind':payload_json['predicted_bridged_ind']}]}]
+        {
+            "inputDataSource": f"{claim_number}:0",
+            "entityId": claim_number,
+            "predictedResult": [
+                {
+                    "claimNumber": claim_number,
+                    "predictedProbability": payload_json["predicted_probability"],
+                    "predicted_bridged_ind": payload_json["predicted_bridged_ind"],
+                }
+            ],
+        }
+    ]
 
 
 # print(
@@ -275,4 +350,3 @@ def predict(**kwargs):
 #         },
 #     )
 # )
-
